pyproject.toml
[project]
name = "3candles"
version = "0.1.0"
description = ""
authors = [
    {name = "gvardovski",email = "gvardovski@icloud.com"}
]
readme = "README.md"
requires-python = "3.13.3"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "vectorbt (>=0.28.1,<0.29.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pyaml (>=25.7.0,<26.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
